third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/sibling-rules-dynamic-changes-expected.txt

CONSOLE ERROR: line 19: Uncaught TypeError: host.createShadowRoot is not a function
Test dynamically changing sibling rule matches in ShadowRoots

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Dynamically change the :first-child
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
